Revert "Factor out utils from compile and vyper-compile tests"
This reverts commit 0e6d5c89302c46cd2d519423c632a0983f0bbf2a.
